LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-navigation-cycles-expected.txt



FAIL CSS Test: ::scroll-marker key navigation cycles assert_equals: first ::scroll-marker gets activated upon clicking expected "rgb(0, 128, 0)" but got ""
